Move v2 runtime handlers out of the v2 package
