+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/LinuxRuntimeHandlerImplV2.java
@@ -4,14 +4,13 @@
  * license information.
  */
 
-package com.microsoft.azure.maven.webapp.handlers.v2;
+package com.microsoft.azure.maven.webapp.handlers;
 
 import com.microsoft.azure.management.appservice.AppServicePlan;
 import com.microsoft.azure.management.appservice.OperatingSystem;
 import com.microsoft.azure.management.appservice.WebApp;
 import com.microsoft.azure.management.appservice.WebApp.Update;
 import com.microsoft.azure.maven.webapp.WebAppUtils;
-import com.microsoft.azure.maven.webapp.handlers.BaseRuntimeHandler;
 
 public class LinuxRuntimeHandlerImplV2 extends BaseRuntimeHandler {
     public static class Builder extends BaseRuntimeHandler.Builder<Builder>{
