+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/PrivateDockerHubRuntimeHandlerImplV2.java
@@ -4,33 +4,31 @@
  * license information.
  */
 
-package com.microsoft.azure.maven.webapp.handlers.v2;
+package com.microsoft.azure.maven.webapp.handlers;
 
 import com.microsoft.azure.management.appservice.AppServicePlan;
 import com.microsoft.azure.management.appservice.OperatingSystem;
 import com.microsoft.azure.management.appservice.WebApp;
 import com.microsoft.azure.maven.Utils;
 import com.microsoft.azure.maven.webapp.WebAppUtils;
-import com.microsoft.azure.maven.webapp.handlers.BaseRuntimeHandler;
 import org.apache.maven.settings.Server;
 
 import static com.microsoft.azure.maven.Utils.assureServerExist;
 
-public class PrivateRegistryRuntimeHandlerImplV2 extends BaseRuntimeHandler {
+public class PrivateDockerHubRuntimeHandlerImplV2 extends BaseRuntimeHandler {
     public static class Builder extends BaseRuntimeHandler.Builder<Builder>{
-
         @Override
-        protected PrivateRegistryRuntimeHandlerImplV2.Builder self() {
+        protected PrivateDockerHubRuntimeHandlerImplV2.Builder self() {
             return this;
         }
 
         @Override
-        public PrivateRegistryRuntimeHandlerImplV2 build() {
-            return new PrivateRegistryRuntimeHandlerImplV2(this);
+        public PrivateDockerHubRuntimeHandlerImplV2 build() {
+            return new PrivateDockerHubRuntimeHandlerImplV2(this);
         }
     }
 
-    private PrivateRegistryRuntimeHandlerImplV2(final Builder builder) {
+    private PrivateDockerHubRuntimeHandlerImplV2(final Builder builder) {
         super(builder);
     }
 
@@ -42,7 +40,7 @@ public class PrivateRegistryRuntimeHandlerImplV2 extends BaseRuntimeHandler {
         final AppServicePlan plan = WebAppUtils.createOrGetAppServicePlan(servicePlanName, resourceGroup, azure,
             servicePlanResourceGroup, region, pricingTier, log, OperatingSystem.LINUX);
         return WebAppUtils.defineLinuxApp(resourceGroup, appName, azure, plan)
-            .withPrivateRegistryImage(image, registryUrl)
+            .withPrivateDockerHubImage(image)
             .withCredentials(server.getUsername(), server.getPassword());
     }
 
@@ -54,7 +52,7 @@ public class PrivateRegistryRuntimeHandlerImplV2 extends BaseRuntimeHandler {
         final Server server = Utils.getServer(settings, serverId);
         assureServerExist(server, serverId);
         return app.update()
-            .withPrivateRegistryImage(image, registryUrl)
+            .withPrivateDockerHubImage(image)
             .withCredentials(server.getUsername(), server.getPassword());
     }
 }
