+++ azure-webapp-maven-plugin/src/test/java/com/microsoft/azure/maven/webapp/handlers/v2/LinuxRuntimeHandlerImplV2Test.java
@@ -12,6 +12,7 @@ import com.microsoft.azure.management.appservice.implementation.SiteInner;
 import com.microsoft.azure.maven.auth.AzureAuthFailureException;
 import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.configuration.RuntimeSetting;
+import com.microsoft.azure.maven.webapp.handlers.LinuxRuntimeHandlerImplV2;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.junit.Before;
 import org.junit.Test;
