+++ azure-webapp-maven-plugin/src/test/java/com/microsoft/azure/maven/webapp/handlers/v2/PrivateDockerHubRuntimeHandlerImplV2Test.java
@@ -11,6 +11,7 @@ import com.microsoft.azure.management.appservice.implementation.SiteInner;
 import com.microsoft.azure.maven.auth.AzureAuthFailureException;
 import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.configuration.RuntimeSetting;
+import com.microsoft.azure.maven.webapp.handlers.PrivateDockerHubRuntimeHandlerImplV2;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.settings.Server;
 import org.apache.maven.settings.Settings;
