+++ azure-webapp-maven-plugin/src/test/java/com/microsoft/azure/maven/webapp/handlers/v2/PublicDockerHubRuntimeHandlerImplV2Test.java
@@ -11,6 +11,7 @@ import com.microsoft.azure.management.appservice.implementation.SiteInner;
 import com.microsoft.azure.maven.auth.AzureAuthFailureException;
 import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.configuration.RuntimeSetting;
+import com.microsoft.azure.maven.webapp.handlers.PublicDockerHubRuntimeHandlerImplV2;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.junit.Before;
 import org.junit.Test;
